feat: merged SystemUtils and ImageUtils into one Utils module
