+++ Caching/Package.swift
@@ -15,9 +15,9 @@ let package = Package(
     dependencies: [
         .package(
             url: "https://github.com/Swiftgram/TDLibKit.git",
-            .revisionItem("e78aef926cdd92323755bcb72309aa5afed1f02a")
+            .exact("1.2.1-tdlib-1.8.1-1e1ab5d1")
         ),
-        .package(url: "https://github.com/hyperoslo/Cache.git", .exactItem(.init(6, 0, 0)))
+        .package(url: "https://github.com/hyperoslo/Cache.git", .upToNextMajor(from: "6.0.0"))
     ],
     targets: [.target(
             name: "Caching",
