--- ImageUtils/README.md
@@ -1,5 +1,0 @@
-# ImageUtils
-
-A module of Moc that contains utilities specific to
-image handling and convenient extensions for existing
-types.
