+++ Moc/Views/Preferences/Tabs/AccountsPrefView.swift
@@ -8,13 +8,12 @@
 import AlertToast
 import Backend
 import Combine
-import ImageUtils
+import Utils
 import Logging
 import Resolver
 import SPSafeSymbols
 import SwiftUI
 import SwiftUIUtils
-import SystemUtils
 import TDLibKit
 
 struct AccountsPrefView: View {
