--- SystemUtils/README.md
@@ -1,3 +1,0 @@
-# SystemUtils
-
-This module contains all useful utilities related to system stuff.
