+++ Utils/Tests/UtilsTests/UtilsTests.swift
@@ -1,0 +1,11 @@
+import XCTest
+@testable import Utils
+
+final class UtilsTests: XCTestCase {
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
+        XCTAssertEqual(Utils().text, "Hello, World!")
+    }
+}
